[frontend] New Education Page
* Changed the display of courses
* Changed the carousel to use the shadcn/ui carousel component instead of own implementation
* Adjusted CSS of the static component, both the <Head /> and <Section />
