--- frontend/src/app/[language]/education/box.css
@@ -1,3 +1,0 @@
-.highlight:hover > * {
-    height: 100%;
-}
